+++ packages/sdk/src/merkleProver/index.ts
@@ -1,3 +1,0 @@
-export { MerkleProver } from "./abstract";
-export { InMemoryMerkleProver } from "./inMemory";
-export { MockMerkleProver } from "./mock";